+++ cpp/src/async/signaling.cpp
@@ -12,26 +12,26 @@ class SignalStore {
     SignalStore() = default; 
 
     // create a new signal and store it, returning the signal id 
-    size_t New() {
+    int New() {
         std::unique_lock _l1 {m_storeLock};
         m_lastSignalID += 1;
         m_signalStore.insert(m_lastSignalID);
         return m_lastSignalID;
     }
     // store a signal 
-    void Store(size_t value) {
+    void Store(int value) {
         std::unique_lock _l1 {m_storeLock}; 
         m_signalStore.insert(value);
     }; 
     // check if a signal exists
-    bool Check(size_t value) {
+    bool Check(int value) {
         std::unique_lock _l1 {m_storeLock}; 
         auto loc = m_signalStore.find(value);
         if (loc == m_signalStore.end()) return false; 
         else return true; 
     }
     // consume a signal (check then remove the signal from store)
-    bool Consume(size_t value) {
+    bool Consume(int value) {
         std::unique_lock _l1 {m_storeLock}; 
         auto loc = m_signalStore.find(value);
         if (loc == m_signalStore.end()) return false; 
@@ -41,8 +41,8 @@ class SignalStore {
 
     private:
     std::mutex m_storeLock; 
-    size_t m_lastSignalID = 0; // protected by m_storeLock
-    boost::container::set<size_t> m_signalStore; // protected by m_storeLock
+    int m_lastSignalID = 0; // protected by m_storeLock
+    boost::container::set<int> m_signalStore; // protected by m_storeLock
 };
 
 }
